chore: reorganize project file structure
