+++ src/App.js
@@ -1,4 +1,4 @@
-import GlobalStyle from './styles/globalStyles';
+import GlobalStyle from './components/GlobalStyles/globalStyles';
 import {AppRouter} from './routes/AppRouter';
 
 function App() {
