+++ src/routes/AppRouter.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import {BrowserRouter as Router, Switch, Route} from 'react-router-dom';
-import {Products} from '../pages/Products';
-import {Cart} from '../pages/Cart';
-import {Home} from '../pages/Home';
+import {Products} from '../pages/Products/Products';
+import {Cart} from '../pages/Cart/Cart';
+import {Home} from '../pages/Home/Home';
+import { Login } from '../pages/Login/Login';
 import {Header} from '../components/Header';
-import { Login } from '../pages/Login';
 
 export const AppRouter = () => {
   return (
